notes_bot/requirements.txt
python-telegram-bot>=20.0
psycopg2-binary~=2.9
redis>=7.0
pydantic_settings>=2.11.0
kafka-python>=2.3.0